adding option to run locally without needing database

+++ main.py
@@ -37,17 +37,18 @@ def get_payload_for_post_req():
 
     return payload
 
-def check_for_job_and_send_alert():
+def check_for_job_and_send_alert(use_db, local_job_array):
     payload = get_payload_for_post_req()
     response = requests.post(BASE_URL, data=payload)
     soup = BeautifulSoup(response.content, "html.parser")
     listings = soup.find_all("div", attrs={"class": "listing row"})
 
-    password = quote_plus(DB_SECRET)
-    uri = f"mongodb+srv://karpit:{password}@cluster0.xi7lz9a.mongodb.net/?retryWrites=true&w=majority"
-    client = MongoClient(uri)
-    db = client["job_alert_amzn"]
-    collection = db["alert"]
+    if use_db:
+        password = quote_plus(DB_SECRET)
+        uri = f"mongodb+srv://karpit:{password}@cluster0.xi7lz9a.mongodb.net/?retryWrites=true&w=majority"
+        client = MongoClient(uri)
+        db = client["job_alert_amzn"]
+        collection = db["alert"]
 
     if listings:
 
@@ -55,34 +56,52 @@ def check_for_job_and_send_alert():
         for listing in listings:
             current_job_ids.append(listing.find("strong").text)
         
-        data = list(collection.find())
+        if use_db:
+            data = list(collection.find())
 
-        for dic in list(data):
-            if dic["job_id"] not in current_job_ids:
-                collection.delete_one({"job_id": dic["job_id"]})
-                data.remove(dic)      
-        
+            for dic in list(data):
+                if dic["job_id"] not in current_job_ids:
+                    collection.delete_one({"job_id": dic["job_id"]})
+                    data.remove(dic)
+        else:
+            for dic in list(local_job_array):
+                if dic["job_id"] not in current_job_ids:
+                    local_job_array.remove(dic)
+
         for listing in listings:
 
             job_desc = listing.find("a").text
             job_location = listing.find("span").text
             job_id = listing.find("strong").text
             job_link = BASE_URL.replace("/BBIndex", "") + listing.find("a")["href"]
+            
+            if use_db:
+                if not any(dic["job_id"] == job_id for dic in data):
+                    collection.insert_one({"job_id": job_id,
+                                        "job_desc": job_desc,
+                                        "location": job_location,
+                                        "link": job_link
+                                        })                
+                    send_alert(job_location, job_desc, job_link)
+            else:
+                if not any(dic["job_id"] == job_id for dic in local_job_array):
+                    local_job_array.append({"job_id": job_id,
+                                        "job_desc": job_desc,
+                                        "location": job_location,
+                                        "link": job_link
+                                        })                
+                    send_alert(job_location, job_desc, job_link)
 
-            if not any(dic["job_id"] == job_id for dic in data):
-                collection.insert_one({"job_id": job_id,
-                                    "job_desc": job_desc,
-                                    "location": job_location,
-                                    "link": job_link
-                                    })                
-                send_alert(job_location, job_desc, job_link)
-
-        client.close()
+        if use_db:
+            client.close()
     else:
-        collection.delete_many({})
+        if use_db:
+            collection.delete_many({})
+        else:
+            local_job_array = []
 
 def main():
-    check_for_job_and_send_alert()
+    check_for_job_and_send_alert(use_db=True, local_job_array=[])
         
 if __name__ == "__main__":
     main()
